Fixed issues detected by ShellCheck

+++ 01-quality_control/02-generate_multiqc_report_pe.zsh
@@ -25,13 +25,13 @@ then
     DATASETS=(
         "2015-jasper_worker_tissues"
         )
-    echo "Processing honey bee data sets: ${DATASETS}"
+    echo "Processing honey bee data sets: " "${DATASETS[@]}"
 else
     DATASETS=(
         "2017-manfredini_queen_brain"
         "2019-porath_worker_brain"
         )
-    echo "Processing bumble bee data sets: ${DATASETS}"
+    echo "Processing bumble bee data sets: " "${DATASETS[@]}"
 fi
 
 # Create symbolic links for "input", "softw", and "tmp" directories
@@ -40,23 +40,23 @@ ln -sfn ../softw .
 ln -sfn ../tmp .
 
 # Run FastQC, STAR, and Qualimap
-for dataset in ${DATASETS}; do
-    mkdir ${dataset}
-    cd ${dataset}
+for dataset in "${DATASETS[@]}"; do
+    mkdir "${dataset}"
+    cd "${dataset}" || exit
 
     # Run FastQC
-    echo "Running FastQC with data set "${dataset}
+    echo "Running FastQC with data set ""${dataset}"
     mkdir ${DATE}-fastqc
-    cd ${DATE}-fastqc
-    parallel -j 8 'fastqc {}' ::: ../../input/${dataset}/*.fastq.gz > fastqc.log 2>&1
-    mv ../../input/${dataset}/*.{zip,html} .
+    cd ${DATE}-fastqc || exit
+    parallel -j 8 'fastqc {}' ::: ../../input/"${dataset}"/*.fastq.gz > fastqc.log 2>&1
+    mv ../../input/"${dataset}"/*.{zip,html} .
     echo "Finished running FastQC"
     cd ..
 
     # Run STAR
-    echo "Running STAR with data set "${dataset}
+    echo "Running STAR with data set ""${dataset}"
     mkdir ${DATE}-star
-    cd ${DATE}-star
+    cd ${DATE}-star || exit
     
     if [ ${SPECIES} = "Amel" ]
     then
@@ -70,10 +70,10 @@ for dataset in ${DATASETS}; do
     mkdir ${SCRATCH}/${DATE}
     ln -s ${SCRATCH}/${DATE} tmp
     mkdir -p tmp/star
-    for file in ../../input/${dataset}/*_1.fastq.gz; do
-        identifier=$(basename $file _1.fastq.gz)
+    for file in ../../input/"${dataset}"/*_1.fastq.gz; do
+        identifier=$(basename "$file" _1.fastq.gz)
         echo "Writing STAR commands for sample ${identifier}"
-        output=tmp/star/$(basename $file _1.fastq.gz)
+        output=tmp/star/$(basename "$file" _1.fastq.gz)
         echo "STAR --runThreadN 2 --genomeDir ./star_index/ --readFilesCommand zcat \
         --outFileNamePrefix=${output}. --outTmpDir=${output} \
         --readFilesIn ../../input/${dataset}/${identifier}"_1.fastq.gz" ../../input/${dataset}/${identifier}"_2.fastq.gz"" >> \
@@ -83,7 +83,7 @@ for dataset in ${DATASETS}; do
     echo "Finished running STAR commands"
 
     # Convert SAM files to sorted BAM
-    echo "Creating BAM files with data set "${dataset}
+    echo "Creating BAM files with data set ""${dataset}"
     parallel -j 8 'samtools view -@ 4 -bS {} | \
     samtools sort -@ 4 - -o {.}.sorted.bam' ::: tmp/star/*.sam > samtools.log 2>&1
     echo "Finished creating BAM files"
@@ -91,13 +91,13 @@ for dataset in ${DATASETS}; do
     cd ..
 
     # Run Qualimap
-    echo "Running Qualimap with data set "${dataset}
+    echo "Running Qualimap with data set ""${dataset}"
     module unload fastqc/0.11.9
     module unload java
     module load qualimap/2.2.1
 
     mkdir ${DATE}-qualimap
-    cd ${DATE}-qualimap
+    cd ${DATE}-qualimap || exit
     ln -s ${SCRATCH}/${DATE}/star star
     if [ ${SPECIES} = "Amel" ]
     then
@@ -116,9 +116,9 @@ for dataset in ${DATASETS}; do
     cd ..
 
     # Run kallisto
-    echo "Running kallisto with data set "${dataset}
+    echo "Running kallisto with data set ""${dataset}"
     mkdir ${DATE}-kallisto
-    cd ${DATE}-kallisto
+    cd ${DATE}-kallisto || exit
     ln -s ../../softw/kallisto/kallisto kallisto
     ln -s ${SCRATCH}/${DATE} tmp
     mkdir -p tmp/kallisto
@@ -132,8 +132,8 @@ for dataset in ${DATASETS}; do
         ln -s ../../tmp/Bter_kallisto_index/Bter_kallisto.idx kallisto_index
     fi
 
-    for file in ../../input/${dataset}/*_1.fastq.gz; do
-        identifier=$(basename $file _1.fastq.gz)
+    for file in ../../input/"${dataset}"/*_1.fastq.gz; do
+        identifier=$(basename "$file" _1.fastq.gz)
         echo "Writing kallisto commands for sample ${identifier}"
         output=tmp/kallisto/${identifier}
         echo "./kallisto quant -i kallisto_index --bias --bootstrap-sample=100 \
@@ -151,18 +151,18 @@ for dataset in ${DATASETS}; do
     # Delete BAM files in STAR results
     rm ${DATE}-star/tmp/star/*.bam
     # Move other STAR output files to STAR results directory
-    cd ${DATE}-star
+    cd ${DATE}-star || exit
     mkdir results
     mv tmp/star/* results/
     # Remove temporary directories
-    rm -r tmp ${SCRATCH}/${DATE}
+    rm -r tmp ${SCRATCH:?}/${DATE}
     cd ..
 
     # Generate MultiQC report
     multiqc .
     # Rename MultiQC report and data
-    mv multiqc_report.html ${dataset}-multiqc_report.html
-    mv multiqc_data ${dataset}-multiqc_data
+    mv multiqc_report.html "${dataset}"-multiqc_report.html
+    mv multiqc_data "${dataset}"-multiqc_data
     cd ..
 
     module unload qualimap/2.2.1
